changing implementation to use struct fields to hold the methods
due to the fact that interface receivers are not allowed, there is no way
to make the signature of the Func field of a Node generic. To work around
this, we now give the JSON config processor an anonymous struct, which
contains all the fuctions as just plain functions rather than methods.